main2.goを合体しました
