--- go/main2.go
@@ -1,36 +1,0 @@
-import (
-	"database/sql"
-	_ "github.com/mattn/go-sqlite3"
-)
-
-var db *sql.DB
-
-func init() {
-	var err error
-	db, err = sql.Open("sqlite3", "mercari.sqlite3")
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-
-func itemsHandler(w http.ResponseWriter, r *http.Request) {
-	rows, err := db.Query("SELECT id, name, category, image_name FROM items")
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer rows.Close()
-
-	var items []Item
-	for rows.Next() {
-		var item Item
-		if err := rows.Scan(&item.ID, &item.Name, &item.Category, &item.ImageName); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		items = append(items, item)
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(items)
-}
